backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg[binary]>=3.1,<4.0
celery>=5.4,<6.0
redis>=5.0,<6.0
django-environ>=0.11,<1.0
gunicorn>=22.0,<23.0
minio>=7.2,<8.0
pgvector>=0.4,<1.0
whitenoise>=6.8,<7.0
django-cors-headers>=4.6,<5.0

